+++ src/startup.rs
@@ -1,0 +1,15 @@
+use actix_web::{App, HttpServer, web, dev::Server};
+use std::net::TcpListener;
+
+use crate::routes::{health_check, subscribe};
+
+pub fn run(listener: TcpListener) -> Result<Server, std::io::Error> {
+    let server = HttpServer::new(|| {
+        App::new()
+            .route("/health_check", web::get().to(health_check))
+            .route("/subscriptions", web::post().to(subscribe))
+    })
+    .listen(listener)?
+    .run();
+    Ok(server)
+}
